Revert "Correctly invalidate session for SAML Remote Logout (#5990)"
This reverts commit 0625bb871908bddb1906f5a686f36b6b6acb477b.

[skip changelog]
