+++ app/controllers/concerns/saml_idp_logout_concern.rb
@@ -22,9 +22,8 @@ module SamlIdpLogoutConcern
   end
 
   def handle_valid_sp_remote_logout_request(user_id)
-    # Remotely delete the user's current session
-    session_id = User.find(user_id).unique_session_id
-    OutOfBandSessionAccessor.new(session_id).destroy
+    # Remotely invalidate the user's current session, see config/initializers/session_limitable.rb
+    User.find(user_id).update!(unique_session_id: nil)
 
     # rubocop:disable Rails/RenderInline
     render inline: logout_response, content_type: 'text/xml'
